Move pageHeader from pages directory to components directory
